+++ test/structures/MarginalScalingModel/generalscaling_test.jl
@@ -70,7 +70,12 @@
         pd = GeneralScaling(60, 100, 1, 0, 0.8, 5)
         buffer = IOBuffer()
         @test_logs Base.show(buffer, pd)
+    end
 
+    @testset "_neutral_initial_model" begin
+        model = GeneralScaling(1., 20. ,5., .1, .8, .5)
+        neutral_model = GeneralScaling(1., 20. ,5., 0., .8, .5)
+        @test neutral_model == IDFCurves._neutral_initial_model(model)
     end
 
     @testset "cdf(::GeneralScaling)" begin
@@ -83,7 +88,8 @@
     @testset "loglikelihood(::GeneralScaling)" begin
 
         pd = GeneralScaling(3, 1, 1, 0.1, 0.5, 1)
-        data = rand(pd, [1, 3], 3, tags=["1", "3"])
+        duration_dict = Dict(zip(["1", "3"], [1, 3]))
+        data = rand(pd, duration_dict)
         y₁ = getdata(data, "1")
         y₃ = getdata(data, "3")
 
@@ -106,7 +112,8 @@
         n = 1
         d = [0.5, 1, 24]
         tag = ["1", "2", "3"]
-        data = rand(pd, d)
+        duration_dict = Dict(zip(tag, d))
+        data = rand(pd, duration_dict)
 
         @test issetequal(gettag(data), tag)
         for i in eachindex(tag)
@@ -116,27 +123,12 @@
         end
 
         n = 3
-        d = [0.5, 1, 24]
-        tag = ["1", "2", "3"]
-        data = rand(pd, d, n)
+        data = rand(pd, duration_dict, n)
 
         @test issetequal(gettag(data), tag)
         for i in eachindex(tag)
             @test getduration(data, tag[i]) ≈ d[i]
             @test getyear(data, tag[i]) == collect(1:n)
-            @test length(getdata(data, tag[i])) == n
-        end
-
-        n = 3
-        d = [0.5, 1, 24]
-        tag = ["10", "11", "12"]
-        x = [10, 11, 12]
-        data = rand(pd, d, n, tags=tag, x=[10, 11, 12])
-
-        @test issetequal(gettag(data), tag)
-        for i in eachindex(tag)
-            @test getduration(data, tag[i]) ≈ d[i]
-            @test getyear(data, tag[i]) == x
             @test length(getdata(data, tag[i])) == n
         end