+++ test/structures/MarginalScalingModel/simplescaling_test.jl
@@ -53,6 +53,12 @@
 
     end
 
+    @testset "_neutral_initial_model" begin
+        model = SimpleScaling(1., 20. ,5., .1, .8)
+        neutral_model = SimpleScaling(1., 20. ,5., 0., .8)
+        @test neutral_model == IDFCurves._neutral_initial_model(model)
+    end
+
     @testset "cdf(::SimpleScaling)" begin
         pd = SimpleScaling(1, 100, 1, 0, 0.8)
 
@@ -63,7 +69,8 @@
     @testset "loglikelihood(::SimpleScaling)" begin
 
         pd = SimpleScaling(4, 2, 1, -0.1, 0.5)
-        data = rand(pd, [1, 4], 3, tags=["1", "4"])
+        duration_dict = Dict(zip(["1", "4"], [1, 4]))
+        data = rand(pd, duration_dict, 3)
         y₁ = getdata(data, "1")
         y₃ = getdata(data, "4")
 
@@ -86,7 +93,8 @@
         n = 1
         d = [0.5, 1, 4]
         tag = ["1", "2", "3"]
-        data = rand(pd, d)
+        duration_dict = Dict(zip(tag, d))
+        data = rand(pd, duration_dict)
 
         @test issetequal(gettag(data), tag)
         for i in eachindex(tag)
@@ -96,27 +104,12 @@
         end
 
         n = 3
-        d = [0.5, 1, 4]
-        tag = ["1", "2", "3"]
-        data = rand(pd, d, n)
+        data = rand(pd, duration_dict, n)
 
         @test issetequal(gettag(data), tag)
         for i in eachindex(tag)
             @test getduration(data, tag[i]) ≈ d[i]
             @test getyear(data, tag[i]) == collect(1:n)
-            @test length(getdata(data, tag[i])) == n
-        end
-
-        n = 3
-        d = [0.5, 1, 4]
-        tag = ["10", "11", "12"]
-        x = [10, 11, 12]
-        data = rand(pd, d, n, tags=tag, x=[10, 11, 12])
-
-        @test issetequal(gettag(data), tag)
-        for i in eachindex(tag)
-            @test getduration(data, tag[i]) ≈ d[i]
-            @test getyear(data, tag[i]) == x
             @test length(getdata(data, tag[i])) == n
         end